Functional generator of random or connected random graphs
If the option CONNECTED is selected, it generates a random connected graph without an assurence of existing Eulerian trail or cycle. At first, it generates a random spanning tree, then it adds other random edges.
If the option RANDOM is selected, it generates a general random graph with the possibility of isolated vertices.
Other options are just in development.

The graph is represented by matrix (internally by one-dimensional array) and the matrix is printed to the output file or stdout, according to which option is selected.

Tested on eva (FreeBSD).