Edited function for generating graphs with Eulerian trails
The function is now idiot-proof. The input with illegal combination of nodes and vertices is handled - the generating algorithm will finish sooner, even if there are still some remaining edges. Hovewer, in some cases it can finish sooner although whether there was legal input. And in some cases, it generates graphs with Eulerian cycles.

The matrix diagonale does not contain zeroes now. Every element of the diagonale is used as degree counter for the specific node (row / column).