Added function for generating graphs with Eulerian cycles
If the option EULER_CYCLE is selected, it will try to generate a random connected graph with an Eulerian cycle. It is not ensured, that the resulting graph will contain all the set edges and an Eulerian cycle. In some cases, it generates graph just with an Eulerian trail, instead of Eulerian cycle.

The matrix diagonale does not contain zeroes. Every element of the diagonale is used as degree counter for the specific node (row / column).